// src/interfaces/bike.ts
export default interface Bike {
  date_stolen: number;
  description: string;
  frame_colors: string[];
  frame_model: string;
  id: number;
  is_stock_img: boolean;
  large_img: string;
  location_found: string | null;
  manufacturer_name: string;
  external_id: string | null;
  registry_name: string | null;
  registry_url: string | null;
  serial: string;
  status: string;
  stolen: boolean;
  stolen_coordinates: [number, number];
  stolen_location: string;
  thumb: string;
  title: string;
  url: string;
  year: number;
  propulsion_type_slug: string;
  cycle_type_slug: string;
  registration_created_at: number;
  registration_updated_at: number;
  api_url: string;
  manufacturer_id: number;
  paint_description: string | null;
  name: string;
  frame_size: string;
  rear_tire_narrow: boolean;
  front_tire_narrow: boolean | null;
  type_of_cycle: string;
  test_bike: boolean;
  rear_wheel_size_iso_bsd: number;
  front_wheel_size_iso_bsd: number;
  handlebar_type_slug: string | null;
  frame_material_slug: string;
  front_gear_type_slug: string | null;
  rear_gear_type_slug: string | null;
  extra_registration_number: string | null;
  additional_registration: string | null;
  stolen_record: {
    date_stolen: number;
    location: string;
    latitude: number;
    longitude: number;
    theft_description: string;
    locking_description: string;
    lock_defeat_description: string;
    police_report_number: string;
    police_report_department: string;
    created_at: number;
    create_open311: boolean;
    id: number;
  };
  public_images: {
    name: string;
    full: string;
    large: string;
    medium: string;
    thumb: string;
    id: number;
  }[];
  components: {
    id: number;
    description: string;
    serial_number: string;
    component_type: string;
    component_group: string;
    rear: null | string;
    front: null | string;
    manufacturer_name: string;
    model_name: string;
    year: null | number;
  }[];
}